refactor: modularize CI summary helpers
