+++ scripts/__tests__/generate-ci-summary.test.js
@@ -6,18 +6,21 @@ import os from 'node:os';
 import { fileURLToPath } from 'node:url';
 import { execFile } from 'node:child_process';
 import { promisify } from 'node:util';
-import { collectTestCases, loadRequirements, mapToRequirements } from '../generate-ci-summary.ts';
+import { collectTestCases } from '../summary/tests.ts';
+import { loadRequirements, mapToRequirements } from '../summary/requirements.ts';
 import { groupToMarkdown, summaryToMarkdown, requirementsSummaryToMarkdown, buildSummary, computeStatusCounts } from '../summary/index.ts';
 import { writeErrorSummary } from '../error-handler.ts';
 
 const fileUrl = new URL('../generate-ci-summary.ts', import.meta.url);
 const summaryUrl = new URL('../summary/index.ts', import.meta.url);
+const requirementsUrl = new URL('../summary/requirements.ts', import.meta.url);
 
 test('generate-ci-summary features', async () => {
   const content = await fs.readFile(fileUrl, 'utf8');
   const summaryContent = await fs.readFile(summaryUrl, 'utf8');
+  const reqContent = await fs.readFile(requirementsUrl, 'utf8');
   assert.match(content, /TEST_RESULTS_GLOBS/);
-  assert.match(content, /<redacted>/);
+  assert.match(reqContent, /<redacted>/);
   assert.match(summaryContent, /<details><summary>/);
   assert.match(content, /\*\*\/\*junit\*\.xml/);
 });
